# openapi.yaml
openapi: 3.0.3
info:
  title: Real Estate Fractional Investment Platform API
  version: 1.0.0
  description: >
    HTTP API for a fractional real estate investment marketplace.
    The client never connects to the database directly; all access
    is mediated through this API. This spec exposes CRUD operations
    for profiles, properties, listings, holdings, transactions,
    notifications, and admin actions.
servers:
  - url: https://api.example.com
    description: Production server
  - url: https://api.sandbox.example.com
    description: Sandbox server
...
paths:
...
  /properties:
    get:
      tags: [Properties]
      summary: List properties
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: search
          description: Search by property name
          schema:
            type: string
      responses:
        '200':
          description: List of properties
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Property'
                  total:
                    type: integer
    post:
      tags: [Properties]
      summary: Create a property
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreate'
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequestError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    BadRequestError:
      description: Bad request – validation or input error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    UUID:
      type: string
      format: uuid
      example: "11111111-1111-1111-1111-111111111111"
    Timestamp:
      type: string
      format: date-time
      example: "2025-01-01T12:00:00Z"
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
      required: [message]
...
    Property:
      type: object
      properties:
        property_id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        location:
          type: string
          nullable: true
        property_type:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
      required:
        - property_id
        - name
    PropertyCreate:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        property_type:
          type: string
        description:
          type: string
      required:
        - name
...